topogdal: read list of rasters from the file containing paths
why? generally paths can be quite long (up to 4096 bytes), and ssdp
script has a limit for the each line. Hence, when one needs to sample
many files the ssdp script is neater when a list of raster paths is
given via a file containing paths in each line.

the flist is an option, and if a file is provided which cannot be
read, the error is raised.