drop 1-array arguments, make documentation names consistent
so far variables can be either a Configuration, Array or a Number.

All Configuration, Array arguments come with in-/out- prefix, to
indicate whether this is an input or an output.

A number **cannot** be output, hence all number are either float-value
or int-value.

Everywhere where it is read length-1 array, I replace with FetchFloat
or FetchInt