fix: incorrect coordinate system is passed to topogrid
this was a tricky one, which goes back to the sampling I implemented
in pvgrip. This bug implies that all our simulations were inaccurate.

Briefly the problem: the coordinate system was slightly inaccurate,
and this inaccuracy is very small in the south-west corner, and the
error grows towards the north-east. The rate of error depends on the
raster size, but for large rasters if observation ends up in the
north-east corner, the error can be up to 20 meters.

Source of error: I made a metric mesh that can be expressed as that is
a produce of a range of latitude and longitude vectors:

        lats = arange(lat_min, lat_max, step)
        lons = arange(lon_min, lon_max, step)

the mistake was that I converted those two vectors back to 4326 system
using:

        convert(lats, lons[0])
        convert(lats[0], lons)

which introduced a bias towards the north-east corner.

The error becomes obvious, when you try to draw something on the
raster and discover that coordinates you select in Openstreetmap are
not the ones you get.

To check that things are correct, assume that `raster_max.list`
contains all the rasters needed for the following script:

        init_sim_config C=C

        make_scalar x=Plat val=50.90368859346916
        make_scalar x=Plon val=6.409964350265447
        make_scalar x=Pz val=100

        make_scalar x=LAT_MIN val=50.898910809991435
        make_scalar x=LON_MIN val=6.405444163123768
        make_scalar x=LAT_MAX val=50.90552355097049
        make_scalar x=LON_MAX val=6.41680950171901
        make_scalar x=STEP val=1

        make_scalar x=DELTA val=0.0021

        determine_utm epsg=EPSG lat=LAT_MIN lon=LON_MIN
        convert_epsg x=Plat y=Plon epsg_src=4326 epsg_dst=EPSG

        config_topogdal C=C lat1=LAT_MIN lon1=LON_MIN lat2=LAT_MAX lon2=LON_MAX step=STEP epsg=EPSG flist=raster_max.list
        addheight_topo C=C y=Plat x=Plon z=Pz
        get_grid C=C x=x y=y nx=nx ny=ny
        sample_topo C=C x=x y=y z=z azimuth=azi zenith=zen
        write_png z=z nx=nx ny=ny ofn=delta0.png normalise=1

        array_eval a=LAT_MIN op=+ b=DELTA c=LAT_MIN
        array_eval a=LAT_MAX op=+ b=DELTA c=LAT_MAX

        config_topogdal C=C lat1=LAT_MIN lon1=LON_MIN lat2=LAT_MAX lon2=LON_MAX step=STEP epsg=EPSG flist=raster_max.list
        addheight_topo C=C y=Plat x=Plon z=Pz
        get_grid C=C x=x y=y nx=nx ny=ny
        sample_topo C=C x=x y=y z=z azimuth=azi zenith=zen
        write_png z=z nx=nx ny=ny ofn=delta1.png normalise=1

The scripts renders two rasters with a 100 meters pole fixed at one
spot. The rasters are rendered with a shifted bounding box. In the
previous sampling implementation the location of the pole changes, and
it depends on the bounding box. Now the position does not change, and
it correspond to the locations in OSM.