horizoncache: store traced horizon inside rtree-index
major changes: now the location keeps just a pointer to the allocated
horizon *H.

        typedef struct location {
        	sky_transfer T;
        	horizon* H; // computed horizons are stored in horizoncache
        	double difftrans; // diffuse light transmission efficiency for a uniform sky
        } location;

The memory management of the allocated horizon is delegated to the

        struct horizoncache {
        		struct rtree *rtree;
        		double xydelta;
        		double zdelta;
        };

The memory for `horizon` is allocated in `horizoncache_get` and freed
when `horizoncache_free` is called.

The `horizoncache_get` either returns a pointer to the horizon
corresponding to location (x,y,z). The location is matched up to the
xydelta (x +- xydelta/2, y +- xydelta/2, z +- zdelta/2). If rtree
does not contain a horizon at the given location, the horizon is
initialised with horizon->zen=NULL value. If memory allocation fail
`horizoncache_get` returns NULL, this error should be processed
downstream.

The `horizoncache_free` iterates over all data contained in rtree
index, and frees horizon there.

The cache contains all computed location until the end of the program,
or until `config_cleanlocations` SSDP routine is called.

Memory footprint for the cache is only 22 MB additional memory
required for 165249 traced locations