fix: memory leak when initialising horizon in parallel
the memory leak happens, when at the initialisation of locations, some
locations points to the same horizon. Then different threads will do
double initialisation for the same horizon pointer, which will not be
freed in the end.

To solve this problem, I use a sort of a hashmap (C->uH), that stores
uniquely all the horizons that needs to be computed. Since pointers in
this hashmap are unique, it is safe to initialise horizon in parallel.