approximate horizon: sample topography using polar Sobol set
This introduces optional arguments `approx_n` and `approx_decay` to
`config_locations`. By default, config_locations keeps the old
behaviour of the horizon tracing (approx_n=-1).

The approximate algorithm uses a two-dimensional Sobol' sequence
covering {(x_i,y_i)}_i \in [0,1] x [0,1] square. The square is turned
into a sample pixel location for the horizon using:

        N * x^approx_decay * exp(1i*2*pi*y_i)

The approx_decay controls the density of points near the location. The
approx_n does not correspond precisely to the number of sample points,
as those location are converted to pixels, which could coincide.

The lookup of pixels is done using char* array, as it is faster than
the hashmap (perhaps, my implementation of the hashmap is crappy). But
this introduced additional memory footprint (79 MB for 4501x4439
raster).

To compute the horizon only pixels from this sequence is used. Such an
approximation is always an underestimation of the horizon. Here are
the performance values in terms of accuracy and speed (computed for
complete raster of 338x333 pixels, 112554 locations):

 | RMSE (% of GHI) | bias (% of GHI) | Q_0.99(diff) (% of GHI) | time  | speedup | nsobol | approx_n | approx_decay |
 |-----------------+-----------------+-------------------------+-------+---------+--------+----------+--------------|
 |              NA |              NA |                      NA | 127 s |      NA |     NA |       NA |         2.25 |
 |              7% |             -5% |                    -28% | 8 s   |   15.88 |    866 |     1000 |         2.25 |
 |              4% |             -2% |                    -22% | 9 s   |   14.11 |   1710 |     2000 |         2.25 |
 |              2% |             -1% |                  -4.65% | 12 s  |   10.58 |   4105 |     5000 |         2.25 |
 |            1.4% |           -0.5% |                  -2.12% | 16 s  |    7.94 |   7882 |    10000 |         2.25 |
 |            0.8% |           -0.2% |                  -0.96% | 23 s  |    5.52 |  15028 |    20000 |         2.25 |
 |            0.3% |          -0.07% |                  -0.42% | 39 s  |    3.26 |  28119 |    40000 |         2.25 |

The default value of approx_decay=2.25 is chosen based on simulation
for a single raster. Some limit theorem is needed to justify the
choice.

For the raster of size 4501x4439 the speed up (1e3 locations, 2 CPUs x
2 threads, old laptop):

 | RMSE (% of GHI) | bias (% of GHI) | Q_0.99(diff) (% of GHI) | time    | speedup | nsobol | approx_n | approx_decay |
 |-----------------+-----------------+-------------------------+---------+---------+--------+----------+--------------|
 |              NA |              NA |                      NA | 154 s   |      NA |     NA |       NA |          2.5 |
 |           10.2% |          -7.95% |                 -31.87% | 10.60 s |   14.52 |    956 |     1000 |          2.5 |
 |            6.3% |          -4.61% |                 -25.22% | 10.95 s |   14.06 |   1900 |     2000 |          2.5 |
 |           2.44% |          -1.87% |                  -5.65% | 16.60 s |    9.27 |   4700 |     5000 |          2.5 |
 |           1.44% |          -0.99% |                  -3.18% | 14.27 s |   10.79 |   9311 |    10000 |          2.5 |
 |           0.67% |          -0.52% |                  -1.88% | 23.04 s |    6.68 |  18394 |    20000 |          2.5 |
 |           0.41% |           -0.3% |                  -1.22% | 19.93 s |    7.72 |  36184 |    40000 |          2.5 |

The quasi random sequence should in general be better than random
sequence (¿What can be said about discrepancy bound for
horizons?). Perhaps a different type of quasi random sequence might
yield better results in terms of accuracy.

p.s. Did you know that of the time of writing Ilya Sobol' still alive?