stcache: sky transfer cache
do not reinitialise sky_transfer data for the same orientation and
albedo. It turns out that POA_Sky_Transfer and POA_Albedo_Transfer
takes 80% of the sky transfer initialisation time, which only depends
on the location orientation and albedo.

The cache benefits most those scenarios, where there are a lot of
common orientations.