{write,read,compute,sim}_sky: save sky to h5 file
I use variable length arrays to read and write pointers. This adds
additional `hvl_t <pointer attribute>` attribute, that points to the
pointer and has the length of the data. All the data necessary to
describe the struct is given in h5t_<struct_name> functions.

When you read the data from h5 file, the memory for those pointers is
allocated by H5Dread, which implies that after you read the data, you
need to make sure that actual struct pointers point to the right place
and you cleaned the space before.

When reading sky from the file the datatype is checked. So you won't
be able to read struct data, if the new version has changed.

compute_sky is only implemented for the cumulative perez sky

the struct is saved without any compression, though the compression
can be added later if needed