provide an arbitrary radial distribution for approximate horizon
the `horizon_sample_distr` command allows to specify the horizon
sampling distribution for each of the configured topogrids

Why? Previous Weibull distribution was only good enough for offset 0
measurements. Depending on topography and offset I reckon there can be
different optimal sampling strategies, hence it is easier just to let
the user decide.

Distributions are given as quantiles of equidistant probabilities
between 0 and 1.