topogdal: make sure the rasters are given in meters
the solution is pretty damn. But among the data we have the rasters
either have en empty string for the vertical units, or metres or feet.

If all fails I assume meters and blame the user for not providing the
proper rasters.

The current implementation allows mixing feet and meters, and I tested
things on 50 random rasters from the 3e5 km^2 we currently work with.