cache optimisations: align locations with topography pixels
here in config_locations I make and ranks array that sorts the
locations according to the topography memory alignment. I
config_locations is called before topogrid is configured, or when
topology is used, the ordering of locations is changed.

Different threads process different locations. My hopes here are that
if nearby locations are close together and will processed together,
and therefore, topography data will end up in the shared cache.