ZenithRange: indices of sky patches corresponding to zenith intervals
sky patches are stored in the following order

            z         a
1  0.00000000 0.0000000
2  0.07306029 0.0000000
3  0.07306029 1.0471976
4  0.07306029 2.0943951
5  0.07306029 3.1415927
6  0.07306029 4.1887902
7  0.07306029 5.2359878
8  0.14612059 0.0000000
9  0.14612059 0.5235988
10 0.14612059 1.0471976
...

ordered first by zenith and then by azimuth.

hence for a given zenith interval one can get indices of the sky
patches. this function to be used in HorizTrans to reduce the loop
iterations, where the zenith interval is computed from min and max
values of the horizon.