earth curvature correction
this commit adds `set_earth_correction` command, allowing to set earth
curvature correction per each configured topogrid raster.

let d be the distance to the point from the observer. Then the
elevation correction due to the curvature is given by:

        R_E (1 - cos(d/R_E)),

where R_E is the Earth radius. The following is a good approximation
of the above:

        d^2 / 2 / R_E